Implemented a basic contact form and about page.
The form sends an email but needs some tidying up.
The code to make it function is all in place but I need to tidy up
meesages that are sent and decide what this should look like in the UI.

I am also considering setting up a database collection to store entries.

The about page is a basic placeholder before I think more about it.